bt2/setup.py.in: improve legibility of our_get_config_vars a bit
This is a bit subjective, but avoid early returns / continue.  I like
early returns when there is clearly one "happy" of "main" path in a
function, for instance for error handling or returning early because a
function is a no-op.  But when handling different significant cases, I
prefer an explicit if/else.  It makes the different code paths stand
out visually.

Change-Id: I484eb4ec7187ce3c889507e31349c3b4461f3fc7
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11764