cpp-common/bt2: add C++ graph bindings
Add C++ bindings to create and manipulate graphs.

Create `bt2::Graph` objects with the `bt2::Graph::create()` static
method.

Add components to graphs with the `bt2::Graph::addComponent()`
methods.  There are overloads of this methods for sources, filters and
sinks, and with and without initialization data.

Connect ports with the `bt2::Graph::connectPorts()` method.  The method
does not (yet) provide access to the connection object returned by the
lib.

Run a graph with `bt2::Graph::run()` or `bt2::Graph::run_once()`.

Change-Id: I92719f9dd6e8c77ded20d1bb4a5ddd2a03e8cf0d
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11806
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>