cpp-common/bt2: `CompClsBridge`: pass init method data to user component constructors
Make `CompClsBridge::init` pass the initialization method data (received
from the lib as a `void *`) down to the user component's constructor.

For convenience, user component classes can specify the type of the data
they expect to receive as a template parameter, sparing them of doing a
static_cast.  This template parameter can be omitted, in which case it
defaults to `void`.

Change-Id: I6b64ea6d75d535d7f2f3b8d559bd2698c186f0ca
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11787
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>