sink.ctf.fs: add support to create LTTng index
The LTTng index provides data to efficiently seek in an CTF file by
denoting the sizes and offsets of the packets. This helps when working
with large traces that are only partially read (like in trace-compass).
The source.ctf.fs component is already capable of reading this index,
however the corresponding sink component lacks support for writing it.

We add support to generate the index file, if explicitly requested by
the user.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>